refact(db): move guild config types to a single file
